requirements.txt
PyQt6>=6.5
openpyxl>=3.1
requests>=2.31
python-docx>=1.1
jieba>=0.42.1
